refactor/#5 Restructuring project structure
